Changed Product.productPrice to double so it's children constructor changed too.

+++ phase1/src/model/products/Bicycle.java
@@ -3,7 +3,7 @@ package model.products;
 public class Bicycle extends Vehicle {
     private BicycleType bicycleType;
 
-    public Bicycle(String productName, int productPrice, int numOfProduct, String manufactureCompany, String bicycleType) {
+    public Bicycle(String productName, double productPrice, int numOfProduct, String manufactureCompany, String bicycleType) {
         super(productName, productPrice, numOfProduct, manufactureCompany);
         this.bicycleType = BicycleType.valueOf(bicycleType);
     }
