+++ phase1/src/model/products/Car.java
@@ -4,7 +4,7 @@ public class Car extends Vehicle {
     private int engineCapacity;
     private boolean isAutomatic;
 
-    public Car( String productName, int productPrice, int numOfProduct, String manufactureCompany, int engineCapacity, boolean isAutomatic) {
+    public Car( String productName, double productPrice, int numOfProduct, String manufactureCompany, int engineCapacity, boolean isAutomatic) {
         super(productName, productPrice, numOfProduct, manufactureCompany);
         this.engineCapacity = engineCapacity;
         this.isAutomatic = isAutomatic;
