+++ phase1/src/model/products/Digital.java
@@ -5,7 +5,7 @@ abstract public class Digital extends Product {
     private String dimensions;
     private static int digitalCounter = 1;
 
-    public Digital(String productName, int productPrice, int numOfProduct, double weight, String dimensions) {
+    public Digital(String productName, double productPrice, int numOfProduct, double weight, String dimensions) {
         super( ProductCategory.Digital , productName, productPrice, numOfProduct);
         this.weight = weight;
         this.dimensions = dimensions;
