+++ phase1/src/model/products/InformationStorageEquipment.java
@@ -3,7 +3,7 @@ package model.products;
 abstract public class InformationStorageEquipment extends Digital {
     private int capacity;
 
-    public InformationStorageEquipment(String productName, int productPrice, int numOfProduct, double weight, String dimensions, int capacity) {
+    public InformationStorageEquipment(String productName, double productPrice, int numOfProduct, double weight, String dimensions, int capacity) {
         super(productName, productPrice, numOfProduct, weight, dimensions);
         this.capacity = capacity;
     }
