+++ phase1/src/model/products/Pencil.java
@@ -3,7 +3,7 @@ package model.products;
 public class Pencil extends Stationery {
     private PencilType pencilType;
 
-    public Pencil(String productName, int productPrice, int numOfProduct, String manufactureCountry, String pencilType) {
+    public Pencil(String productName, double productPrice, int numOfProduct, String manufactureCountry, String pencilType) {
         super(productName, productPrice, numOfProduct, manufactureCountry);
         this.pencilType = PencilType.valueOf(pencilType);
     }
