+++ phase1/src/model/products/PersonalComputer.java
@@ -4,7 +4,7 @@ public class PersonalComputer extends Digital {
     private String CPUModel;
     private int RAMMemory;
 
-    public PersonalComputer(String productName, int productPrice, int numOfProduct, double weight, String dimensions, String CPUModel, int RAMMemory) {
+    public PersonalComputer(String productName, double productPrice, int numOfProduct, double weight, String dimensions, String CPUModel, int RAMMemory) {
         super( productName, productPrice, numOfProduct, weight, dimensions);
         this.CPUModel = CPUModel;
         this.RAMMemory = RAMMemory;
