+++ phase1/src/model/products/SSD.java
@@ -4,7 +4,7 @@ public class SSD extends InformationStorageEquipment {
     private int readSpeed;
     private int writeSpeed;
 
-    public SSD(String productName, int productPrice, int numOfProduct, double weight, String dimensions, int capacity, int readSpeed, int writeSpeed) {
+    public SSD(String productName, double productPrice, int numOfProduct, double weight, String dimensions, int capacity, int readSpeed, int writeSpeed) {
         super( productName, productPrice, numOfProduct, weight, dimensions, capacity);
         this.readSpeed = readSpeed;
         this.writeSpeed = writeSpeed;
